+++ Editor/MenuActions/Geometry/DeleteFaces.cs
@@ -4,7 +4,6 @@ using UnityEngine.ProBuilder.MeshOperations;
 
 namespace UnityEditor.ProBuilder.Actions
 {
-    [MenuActionShortcut(typeof(SceneView), KeyCode.Backspace)]
     sealed class DeleteFaces : MenuAction
     {
         public override ToolbarGroup group