+++ Editor/MenuActions/Geometry/SetPivotToSelection.cs
@@ -15,8 +15,7 @@ namespace UnityEditor.ProBuilder.Actions
         static readonly TooltipContent _tooltip = new TooltipContent
             (
                 "Set Pivot to Center of Selection",
-                @"Moves the pivot point of each mesh to the average of all selected elements positions.  This means the pivot point moves to where-ever the handle currently is.",
-                keyCommandSuper, 'J'
+                @"Moves the pivot point of each mesh to the average of all selected elements positions.  This means the pivot point moves to where-ever the handle currently is."
             );
 
         public override SelectMode validSelectModes